Cudf exchange better logs (#140)
Besides some minor structural improvements, the main change in this PR
are better logs. The existing logs are pretty useless in a multi-driver
situation as they don't specify in which pipeline they occurred. Since
the multi-driver-bug won't be the last, I think we should keep these
logs for the moment.